Added in temperature sensing.
Renamed to "firmware," since this will evolve cleanly enough (it looks like).

This is the nice thing about occam-pi development for firmware: once the plumbing is done (the network), you can improve processes within the API of the process network. Incremental development is much safer in this regard.

Have added temperature, which is just another ADC reading, but it needs to be converted to deg. C before it can be used. Cheated, and jumped up to 32 bits to do the conversion. It would be nice to do this within 16 bits.