Read/Write to external EEPROM
Using the 24LC256. This code (very crude) stores the most recently used address at locations 0,1, and stores bytewise temperature data. When the device is reset, it first reports the previous 10 readings, using the stored address as a starting point.

Lots of work to be done now. This code talks to multiple devices on the TWI bus, stores data, reads it back. Time for paper and pencil work to design what this should actually look like for production use.