Many changes.
I should have made more incremental commits.

The EEPROM module wraps some things up and provides a more process-oriented interface to the EEPROM. Its a start, anyway. Kinda app specific, but not horribly so.

The twi3.module will be rolled back into the distribution (and, ultimately, just called "twi.module").

reset-everything.occ clears the Chronodot and resets the "pointer" in the EEPROM.

f2.occ is me playing more with the firmware. Currently, the last proc was a "sprint" late in the evening on using the Modern Device PIR sensor (http://shop.moderndevice.com/products/pir-sensor). It works. Next step will be to wrap it up in a process and debounce it using the appropriate call into the Plumbing library.

That's everything in this commit. Further cleanup on the firmware will follow, with the goal making it all more process-oriented.