Fixed?
I don't know what was causing problems, actually. There were errors being reported in the external interrupts module... but there was nothing wrong with the module. Things "just worked" when compiled on the Mac, and when I flipped back to the Linux VM, things "just worked" there.

This kinda bothers me. But, I'll move on for now.