Updated code.
This is cleaner; the data packet is now built up in a pipeline, and at each stage of the pipeline new data is added to the record. This eliminates the monolithic data gathering process, and new data can be added to the record (and therefore the pipeline) with minimal modification to the overall program structure. (That is, a new kind of data would just be a new stage in the pipeline, which would not effect any of the other elements of the pipeline.)