Clean, working.
This works. That said, I'm pushing up against the flash limits of the Atmega328. I've now got a lot of library code running around in the background of this application, which is not good. I will end up needing to optimize for space ont his, as I don't yet have code in for the SD card. (Fortunately, it is just serial work, so it won't add much to the program beyond the existing "display" routine.)