Under size again, printing updated.
Although the HH:MM:SS printing is cute, getting rid of seconds from the application saves space. So, we only track HH:MM, and we don't bother pretty-printing it. (That won't be useful in the end anyway.)

Output looks like:
0000,0013,I,0115,0013
0000,0014,C,010b,0013
0000,0014,M,00ff,0013
0000,0014,C,010c,0012
0000,0014,M,010f,0012

Of course, this is HEX, still, so I need to use the dec.int routine...