Fixes, changes.
I've been wrestling with interrupts on the Chronodot. That is now resolved. Because the TWI is so slow (it needs to be moved into C, perhaps?), it is possible for us to miss the interrupt. (What bothers me is that I would think, with our lift into occam-pi, this wouldn't be possible.) So, I need a timeout in an ALT to make sure that we clear the error registers on the Chronodot if we don't hear from it for a while. A watchdog, of sorts...

I'm thinking it mostly has to do with being busy doing serial work, which might be the problem.

Other modules modified as necessary through the firmware implementation process. The firmware looks very, very close to done... but it's too big. We're out of flash. I'm not sure how best to handle that at the moment.