Under size, running nicely.
I had to sacrifice some of the process network to make this work. I also had to sacrifice some of the onboard numerical work. I'm now running right under the wire, but both the VM and the code fit in flash, and run nicely. Next step: rebuild the setup in a prototype container for demonstration.