Fixed #1A.
For some reason, the CTRLZ does not flush the buffer on the OpenLog. I thought it did, based on some document I read somewhere. Instead, it sticks a DEC 26 (HEX #1A) into the datastream. Not at all what we want.

I'll take it out, and let the logger run overnight in the bedroom and see what we get.