Set time separately
This firmware looks good to go. I could modify the network so that we track hours and minutes internally, and just use the tick function of the Chronodot. (That is, get the time once at the start, and then just listen for alarms that indicate when we should increment the number of minutes/hours.) This isn't necessary, and I'm bumping up against address 32,000 again...

It works, and I'll let it collect data for a while.