Trigger LED on storage of data.
Now, alert.led is used whenever data is stored. This is useful, as it provides a
visual indicator as to when a reading has been taken.

The network diagram has been updated to reflect how things are connected. I'm
thinking a tech report (at least) should be written about this sprint.